backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dateutil>=2.8.2
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
psycopg2-binary>=2.9.9
requests>=2.31.0
openpyxl>=3.1.0
reportlab>=4.0.0

